requirements.txt
APScheduler==3.6.3
astropy==4.2
certifi==2020.12.5
cffi==1.14.4
cryptography==3.3.1
decorator==4.4.2
numpy==1.19.5
persiantools==1.5.1
pycparser==2.20
pyerfa==1.7.1.1
python-dotenv==0.15.0
python-telegram-bot==13.1
pytz==2020.5
schedule==0.6.0
six==1.15.0
tornado==6.1
tzlocal==2.1
